+++ commercetools/commercetools-sdk-java-importapi/src/main/java-generated/com/commercetools/importapi/models/productvariants/ProductVariantImport.java
@@ -88,14 +88,6 @@ public interface ProductVariantImport extends ImportResource {
     public List<Asset> getAssets();
 
     /**
-     *  <p>If <code>publish</code> is set to either <code>true</code> or <code>false</code>, both staged and current projections are set to the same value provided by the import data. If <code>publish</code> is not set, the staged projection is set to the provided import data, but the current projection stays unchanged. However, if the import data contains no update, that is, if it matches the staged projection of the existing Product, the import induces no change in the existing Product whether <code>publish</code> is set or not.</p>
-     * @return publish
-     */
-    @Deprecated
-    @JsonProperty("publish")
-    public Boolean getPublish();
-
-    /**
      *  <ul>
      *   <li>Set to <code>false</code> to update both the current and staged projections of the Product with the new Product Variant data.</li>
      *   <li>Leave empty or set to <code>true</code> to only update the staged projection.</li>
@@ -182,13 +174,6 @@ public interface ProductVariantImport extends ImportResource {
     public void setAssets(final List<Asset> assets);
 
     /**
-     *  <p>If <code>publish</code> is set to either <code>true</code> or <code>false</code>, both staged and current projections are set to the same value provided by the import data. If <code>publish</code> is not set, the staged projection is set to the provided import data, but the current projection stays unchanged. However, if the import data contains no update, that is, if it matches the staged projection of the existing Product, the import induces no change in the existing Product whether <code>publish</code> is set or not.</p>
-     * @param publish value to be set
-     */
-    @Deprecated
-    public void setPublish(final Boolean publish);
-
-    /**
      *  <ul>
      *   <li>Set to <code>false</code> to update both the current and staged projections of the Product with the new Product Variant data.</li>
      *   <li>Leave empty or set to <code>true</code> to only update the staged projection.</li>
@@ -226,7 +211,6 @@ public interface ProductVariantImport extends ImportResource {
         instance.setAttributes(template.getAttributes());
         instance.setImages(template.getImages());
         instance.setAssets(template.getAssets());
-        instance.setPublish(template.getPublish());
         instance.setStaged(template.getStaged());
         instance.setProduct(template.getProduct());
         return instance;
@@ -261,7 +245,6 @@ public interface ProductVariantImport extends ImportResource {
                         .map(com.commercetools.importapi.models.common.Asset::deepCopy)
                         .collect(Collectors.toList()))
                 .orElse(null));
-        instance.setPublish(template.getPublish());
         instance.setStaged(template.getStaged());
         instance.setProduct(
             com.commercetools.importapi.models.common.ProductKeyReference.deepCopy(template.getProduct()));